Refactor workspaces\templates-lib\packages\template-user-management\src\templateUserManagement.ts
- Move getLoggedInUser into a seperate module the src/client pages, include isAuthenticated as well
- don't move in terms of import / export - it should work similar as done for getEndpoint
- smae for handleRedirectCallback, create a new module and then call the function in that module which has all the logic.
- same with getCookieSettings
- same with performLogout
- also put operationWithRedirect into the /client module - this one can just move the whole function and then import it where required
