+++ workspaces/templates-lib/packages/template-user-management/src/client/state.ts
@@ -1,0 +1,14 @@
+/**
+ * Shared state variables used across authentication modules
+ */
+
+export let forceLogout = false;
+export let refreshTokenStorage: string | undefined = undefined;
+
+export function setForceLogout(value: boolean) {
+  forceLogout = value;
+}
+
+export function setRefreshTokenStorage(token: string | undefined) {
+  refreshTokenStorage = token;
+}
